Change 'File' to 'ImageHelper'
