+++ Samples/WinMLSamplesGallery/WinMLSamplesGallery/Samples/ImageEffects/ImageEffects.xaml.cs
@@ -521,7 +521,7 @@ namespace WinMLSamplesGallery.Samples
 
         private void OpenButton_Clicked(object sender, RoutedEventArgs e)
         {
-            var file = File.PickImageFiles();
+            var file = ImageHelper.PickImageFiles();
             if (file != null)
             {
                 BasicGridView.SelectedItem = null;
